Add condition for set_vring_enable
According to the vhost user spec:
https://qemu-project.gitlab.io/qemu/interop/vhost-user.html#a-vring-state-description

Rings can be in one of three states:

1. stopped: the back-end must not process the ring at all.

2. started but disabled: the back-end must process the ring without
causing any side effects. For example, for a networking device, in the
disabled state the back-end must not supply any new RX packets, but must
process and discard any TX packets.

3. started and enabled.

If VHOST_USER_F_PROTOCOL_FEATURES has not been negotiated, the ring starts directly
in the enabled state.

So, add the condition for the set_vring_enable, only when the
VHOST_USER_F_PROTOCOL_FEATURES is negotiated, it needs to be called.

Change-Id: I2a363dcd11a51ecb7f739afd6319adfa40d1c1fe
Signed-off-by: Conghui Chen <quic_conghuic@quicinc.com>