UPSTREAM: add --lock-guest-memory-dontneed option
`--lock-guest-memory` uses `mlock` on the guest memory and then
`MADV_REMOVE` to free ranges, but `MADV_REMOVE` cannot be used on locked
memory ranges, so it requires sandboxing to be enabled and only performs
the `MADV_REMOVE` calls from the sandboxed balloon process (memory locks
are implicitly release on fork).

The new `--lock-guest-memory-dontneed` added in this commit uses
`MADV_DONTNEED_LOCKED`, which essentially does the same thing as
`MADV_REMOVE`, but it works correctly on locked ranges and so doesn't
require sandboxing. The downside is that it requires a 5.18 or newer
kernel.

BUG=b:407786138

Bug: 407786138
Test: launch protected microdroid
Flag: EXEMPT ...
Change-Id: Ib8e24d647a97b98abf5b3ba6b0356c9d9999a378