ANDROID: use PUNCH_HOLE instead of DONTNEED_LOCKED to free memory
MADV_DONTNEED_LOCKED doesn't eagerly complete all the steps to free the
memory and so, even though the process' RSS goes down, the /proc/meminfo
memavailable stat doesn't go back up.

Instead, use `fallocate` with `FALLOC_FL_PUNCH_HOLE` on the underlying
file.

This is urgent, so I'm modifying the --lock-guest-memory-dontneed
behavior in Android temporarily and will move it to a separte option in
crosvm upstream once it is proven to work.

I tested using the following command, with a VM that started on boot
enabled, and saw an improvement of 300+ MB in MemAvailable.

    adb reboot && adb wait-for-device root && for i in {0..30}; do adb shell cat /proc/meminfo | grep 'MemAvailable\|Shmem:'; sleep 5; done | tee ~/mem_stats

Bug: 413829005
Test: see above
Change-Id: I872f9051afe165b7e10f8a397792268e158e0e33